+++ spooq/transformer/mapper.py
@@ -52,25 +52,19 @@ class Mapper(Transformer):
         clean, pivot, anonymize, ... the data itself. Please have a look at the
         :py:mod:`spooq.transformer.mapper_custom_data_types` module for more information.
 
-    skip_missing_columns : :any:`bool`, Defaults to False
-        Specifies if the mapping transformation should be skipped if a referenced input
-        column is missing in the provided DataFrame. Only one of `nullify_missing_columns` and `skip_missing_columns`
-        can be set to True. If none of the two is set to True then an exception will be raised in case the input
-        column was not found.
-
-    nullify_missing_columns : :any:`bool`, Defaults to False
-        Specifies if the mapping transformation should use NULL if a referenced input
-        column is missing in the provided DataFrame. Only one of `nullify_missing_columns` and `skip_missing_columns`
-        can be set to True. If none of the two is set to True then an exception will be raised in case the input
-        column was not found.
+    missing_column_handling : :any:`str`, Defaults to 'raise_error'
+        Specifies how to proceed in case a source column does not exist in the source DataFrame:
+            * raise_error (default)
+                Raise an exception
+            * nullify
+                Create source column filled with null
+            * skip
+                Skip the mapping transformation
 
     ignore_ambiguous_columns : :any:`bool`, Defaults to False
         It can happen that the input DataFrame has ambiguous column names (like "Key" vs "key") which will
         raise an exception with Spark when reading. This flag surpresses this exception and skips those affected
         columns.
-
-    ignore_missing_columns : :any:`bool`, Defaults to False
-        DEPRECATED: please use nullify_missing_columns instead!
 
     mode : :any:`str`, Defaults to "replace"
         Defines weather the mapping should fully replace the schema of the input DataFrame or just add to it.
@@ -87,6 +81,13 @@ class Mapper(Transformer):
                 The columns provided in the mapping are added at the beginning of the input schema. If a column already
                 exists in the input DataFrame, its position is kept.
                 => output schema: new columns + input columns
+
+
+    Keyword Arguments
+    -----------------
+    ignore_missing_columns : :any:`bool`, Defaults to False
+        DEPRECATED: please use missing_column_handling instead!
+
 
     Note
     ----
@@ -112,38 +113,39 @@ class Mapper(Transformer):
     Note
     ----
     The available input columns can vary from batch to batch if you use schema inference
-    (f.e. on json data) for the extraction. Ignoring missing columns on the input DataFrame is
-    highly encouraged in this case. Although, if you have tight control over the structure
-    of the extracted DataFrame, setting `ignore_missing_columns` to True is advised
-    as it can uncover typos and bugs.
+    (f.e. on json data) for the extraction. Via the parameter `missing_column_handling` you can specify a strategy on
+    how to handle missing columns on the input DataFrame.
+    It is advised to use the 'raise_error' option as it can uncover typos and bugs.
     """
 
     def __init__(
         self,
         mapping: List[Tuple[Any, Any, Any]],
-        skip_missing_columns: bool = False,
-        nullify_missing_columns: bool = False,
         ignore_ambiguous_columns: bool = False,
-        ignore_missing_columns: bool = False,
+        missing_column_handling: str = "raise_error",
         mode: str = "replace",
+        **kwargs
     ):
         super(Mapper, self).__init__()
-        if ignore_missing_columns:
-            self.logger.warn("Parameter `ignore_missing_columns` is deprecated, use `nullify_missing_columns` instead!")
-            warnings.warn(
-                message="Parameter `ignore_missing_columns` is deprecated, use `nullify_missing_columns` instead!",
-                category=FutureWarning
-            )
         self.mapping = mapping
-        self.skip_missing_columns = skip_missing_columns
-        self.nullify_missing_columns = nullify_missing_columns or ignore_missing_columns
-        if self.nullify_missing_columns and self.skip_missing_columns:
-            raise ValueError(
-                "Only one of the parameters `nullify_missing_columns` (before `ignore_missing_columns`) and "
-                "`skip_missing_columns` can be set to True!"
-            )
+        self.missing_column_handling = missing_column_handling
         self.ignore_ambiguous_columns = ignore_ambiguous_columns
         self.mode = mode
+
+        if "ignore_missing_columns" in kwargs:
+            message = "Parameter `ignore_missing_columns` is deprecated, use `missing_column_handling` instead!"
+            if kwargs["ignore_missing_columns"]:
+                message += "\n`missing_column_handling` was set to `nullify` because you defined " \
+                           "`ignore_missing_columns=True`!"
+                self.missing_column_handling = "nullify"
+            self.logger.warn(message)
+            warnings.warn(message=message, category=FutureWarning)
+
+        if self.missing_column_handling not in ["raise_error", "skip", "nullify"]:
+            raise ValueError("""Only the following values are allowed for `missing_column_handling`: 
+                - raise_error: raise an exception in case the source column is missing
+                - skip: skip transformation in case the source is missing
+                - nullify: set source column to null in case it is missing""")
 
     def transform(self, input_df: DataFrame) -> DataFrame:
         self.logger.info("Generating SQL Select-Expression for Mapping...")
@@ -207,14 +209,14 @@ class Mapper(Transformer):
                 source_column = F.col(source_column)
 
         except AnalysisException as e:
-            if isinstance(source_column, str) and self.skip_missing_columns:
+            if isinstance(source_column, str) and self.missing_column_handling == "skip":
                 self.logger.warn(
-                    f"Missing column ({source_column}) skipped (via skip_missing_columns=True): {e.desc}"
+                    f"Missing column ({str(source_column)}) skipped (via missing_column_handling='skip'): {e.desc}"
                 )
                 return None
-            elif isinstance(source_column, str) and self.nullify_missing_columns:
+            elif isinstance(source_column, str) and self.missing_column_handling == "nullify":
                 self.logger.warn(
-                    f"Missing column ({source_column}) replaced with NULL (via nullify_missing_columns=True): {e.desc}"
+                    f"Missing column ({str(source_column)}) replaced with NULL (via missing_column_handling='nullify'): {e.desc}"
                 )
                 source_column = F.lit(None)
             elif "ambiguous" in e.desc.lower() and self.ignore_ambiguous_columns: